+++ webapp/include/Hooks.h
@@ -39,7 +39,7 @@ APIEXPORT unsigned long long GetWebappTime();
 //Database
 APIEXPORT Database* CreateDatabase(Webapp*);
 APIEXPORT void DestroyDatabase(Webapp*, Database*);
-APIEXPORT Database* GetDatabase(Webapp*, int index);
+APIEXPORT Database* GetDatabase(Webapp*, size_t index);
 APIEXPORT int ConnectDatabase(Database*, int database_type, const char* host, const char* username, const char* password, const char* database);
 APIEXPORT long long ExecString(Database*, webapp_str_t* in);
 APIEXPORT int SelectQuery(Query*);
